requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
torchaudio
pillow
opencv-python
numpy
mediapipe
scikit-image
dlib
